requirements.txt
python-telegram-bot==20.8
google-cloud-firestore==2.14.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==8.0.0